working on new ID based join iterator
